DLT Writer: Test CONTROL_TIME messages
These messages are not specified in the AutoSAR DLT standard R19-11 or earlier, but it is supported by the dlt-viewer.

Ensure that we recognise the `IControlArg` type, and ensure that no service identifier is written. Later, when we write the entire payload, we'll test that the DLT Type is `CONTROL_TIME`.

Issue: