DLT Writer: Write SW Injection responses and requests
The response is a simple empty result, as given in the AutoSAR documentation. The request is similar to a raw payload.

If custom encoders are needed, then the user is expected to provide their own `IControlArg` types (perhaps override) and their own `IControlArgEncoder`. Then they should override `ControlArgEncoder` and register them.

Issue: DOTNET-772