DLT: Modernise IVerboseDltDecoder to return Result<int>
One of the problems observed when implementing the decoders, is that it is easy to forget to check the results, so that if -1 is returned, it subtracts the total length instead of being handled as an error.

The `Result<int>` type makes this impossible, so that it must be accessed, and can only be done so if there was no error.

The second advantage is, we will eventually remove the `DltArgError` type, and replace it with the `DltDecodeException`. The exception class can also be given inner exceptions. So if we have a problem with decoding, we can record all instances, and identify the root cause a little easier. Note, we don't throw the exceptions, so there is no stack trace associated with them.

A measurement on real-word data, over 3GB, shows a slight performance decrease of consistent 300ms (from 14.2s to 14.5s). This is about 1.5% decrease and not expected to be noticeable.

Issue: DOTNET-789