DLT: Modernise `IControlArgDecoder` to return `Result<int>`
Because we now return information about the error in the `Result.FromException<int>(new DltParseException(...))`, we can remove the base class `ControlArgDecoderBase` and the we no longer need to wrap the message in a `ControlDecodeError`.

The `GetLogInfoResponseDecoder` can now say where in the decoder it failed parsing. We no longer need to raise exceptions, and can pass this back to the decoder.

The `SwInjectionRequestDecoder` had a virtual method, where if this returned -1, it was ignored. Now it checks the result (so this technique discovered one potential bug in the implementation).

Issue: DOTNET-791