DltDump: Write using the DltFileTraceEncoder
Rewrite the `DltOutput` class to use the full trace encoder implementation instead of our own in the DltDump App. Exposed the functionality in `DltFileTraceEncoder` to write the standard header. This solved a hidden bug that the session id was written with the wrong byte order [1].

When writing skipped lines, the full trace line is written, not just the encoded text. That means there are now four arguments, and one can apply a better filter in newer versions of dlt-viewer that allow filtering on arguments.

[1] Even though it solved the bug, it was not being used (the skipped lines never had a session identifier), but if it were to be used in the future, it would now be correct. From R2-11 [PRS_Dlt_00091] ⌈The Standard Header and the Extended Header shall be in big endian format (MSB first).⌋

Issue: DOTNET-801