DltDump: Provide option to convert non-verbose to verbose
Implement `--nv-verbose` to convert non-verbose to verbose.

The conversion must be done in the `DltOutput` class, so that filtered and unfiltered data is converted. The sequence of steps is:
- The options gets the setting from the user
- `FilterCommand` puts it in `FilterConfig` and gives to `FilterApp`
- `FilterApp` sets it in `OutputStreamFactory`
- `OutputStreamFactory` gives to `DltOutput` for that output. So this means it is only used when the output format is writing to a DLT file (writing to the console results in this option being ignored)

It isn't done as part of the normal loop of `FilterApp.ProcessInput` because that would result only in non-filtered data being processed. You'll see there the code is `output.WriteLine(line)` in `ProcessInput` which doesn't contain the byte payload, which is only available during the actual decoding.

Issue: DOTNET-801