DLT: Ensure that release builds are strict
There was a minor error that the library should have been strict for the last release build (it provides a warning of the tag is missing in the project). The tag is provided, so it generates the correct version, but the next update will not warn of a missing label.