DLT Decoder: Workaround corrupted verbose with no arguments
It is a common error that the length is correct, but the NOAR payload may be incorrect (it is non-zero when there are zero arguments). In this case, accept the length of the DLT paylaod, and ignore that it is in error.

Print a warning into the logs, but decode the message.

When testing such a payload with WireShark, it decodes the line properly but indicates an error when parsing the argument. When testing with DLTViewer by Covesa, it will always ignore the error and assume the length of the DLT message is correct.

In this simple case, we may be able to increase the probability of decoding DLT messages.

Issue: DOTNET-921