DltDump: Properly check that inputs aren't overwritten
There are two main issues that are fixed, both that prevent proper checks of not overwriting input files:

1. Mapping of input files was done incorrectly, and ignored the scheme of the input file. Now the `file://` scheme is properly interpreted as if it is a file.
2. When filtering is used, the file inputs were not checked that any output files don't overwrite the inputs.

The use of `OutputBase` to know if this is a file or not is incorrect. This is because it can be wrapped with a `FilterOutput` object and as such it was assumed that we weren't writing to a file (where the wrapped object could certainly have been). Instead of the `OutputBase` maintaining the list, the `OutputStreamFactory` should get the list of inputs.

The list of inputs were the URIs, which was mostly the correct, but if `file://` was given, or `tcp://`, then this would still be considered an input file. The change now uses `IInputStream` which has already converted the URI and created the correct element. Then we have the knowledge that this is a file, and the correct file name without having to deal with the URI any more.

When when converting a file, a check will be made regardless of using constraints or filters, and we won't overwrite any of the input files, even if the force flag is given.

Issue: DOTNET-933