DltDump Test: Reduce time some tests take when waiting
Use where possible synchronisation primitives and mocks to check for events. However, not all events can be done by either polling or waiting for events, so many of these timeouts remain. The larger the timeout, the more likely we'll find an error. The slower the computer, the less likely the test will work properly.

About 1s of wait time is removed.

Issue: DOTNET-935