DltDump Test: Fix timeout race condition when testing files
Introduced with "DltDump Test: Reduce time some tests take when waiting" (5d9241a), the timeout test was wrong and would often fail due to a race condition. Fixed to wait for the complete timeout. Also ensures the length of the file is more than zero, as it might write the file (0 bytes) and be checked, followed by the proper length.

Issue: DOTNET-950