DltDump: Correct test cases on machines with different timezones
The PCAP header and DLT storage header have timestamps in the timezone of UTC. The test case compared with a string that had a timezone offset for Central European (summer) Time, that on machines that didn't have this timezone set, the test case would fail.

The test cases are now robuster by specifying the original seconds and programmatically converting it to a string for the local time zone.

Issue: DOTNET-967