DltDump: Handle UDP SocketException on Dispose
When `UdpPacketReceiver.Close()` is called while a task is blocked on `UdpPacketReceiver.ReadAsync`, the exception `SocketExceptionError`, but with a different reason is raised. As this worked earlier, add the new condition.

This also ensures the `UdpPacketReceiver.Dispose()` behaves as expected.

The DltDump app is not affected by this change, only a potential warning to the user. The `UdpPacketReceiver.ReadLineAsync()` is used by `TracePacketReader.GetLineAsync()`, which doesn't handle exceptions.

`TracePacketReader.GetLineAsync()` is used by `FilterApp.ProcessInput()` which captures exceptions, and only prints it. The exception should not occur, as the `reader` variable in `FilterApp.ProcessInput()` shouldn't be closed (and therefore, the would print the `SocketError`).

Issue: DOTNET-991