Test Case: TcpServer should handle exceptions on Dispose
This is an integration test only, and doesn't test production code. It checks the behaviour of the client that it exits gracefully when the server is closed. This functionality still works as expected, so no changes in behaviour that can affect DltDump.

The problem arose in the test server `TcpServer`, when `Dispose` is called it closes `m_Listener`. If the function enters `client = await m_Listener.AcceptTcpClientAsync()` when `m_Listener.CLose()` is called, we get the `SocketException`. So handle it as it is intended to be, just exit the listener/server.

Issue: DOTNET-991